civicwatch/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiofiles==23.2.1
Pillow==10.3.0
transformers==4.41.2
torch==2.3.0
sentencepiece==0.2.0
chromadb==0.5.0
anthropic==0.28.0
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.27.0
